Temporary fix for /stopFires
I found this fix just by fiddling with that part of the script. Unfortunately it's only a temp fix. Maybe Pyth3rEx can make it so the entire fire is put out when you execute this command. 

Making that happen would also fix the problem with fires spreading into in-accessible buildings.